+++ src/engine/engine_derivative.c
@@ -1733,7 +1733,10 @@ void mjd_passive_vel(const mjModel* m, mjData* d) {
   int nv_awake = sleep_filter ? d->nv_awake : nv;
   for (int j = 0; j < nv_awake; j++) {
     int i = sleep_filter ? d->dof_awake_ind[j] : j;
-    d->qDeriv[m->D_rowadr[i] + m->D_diag[i]] -= m->dof_damping[i];
+    mjtNum v = d->qvel[i];
+    const mjtNum* poly = m->dof_dampingpoly + mjNPOLY*i;
+    int adr = m->D_rowadr[i] + m->D_diag[i];
+    d->qDeriv[adr] -= mjd_xPolyForce(m->dof_damping[i], poly, v, mjNPOLY, 1);
   }
 
   // flex edge damping
@@ -1771,7 +1774,8 @@ void mjd_passive_vel(const mjModel* m, mjData* d) {
       if (treenum == 2 && !d->tree_awake[id1] && !d->tree_awake[id2]) continue;
     }
 
-    mjtNum B = -m->tendon_damping[i];
+    mjtNum v = d->ten_velocity[i];
+    mjtNum B = -mjd_xPolyForce(m->tendon_damping[i], m->tendon_dampingpoly+mjNPOLY*i, v, mjNPOLY, 1);
 
     if (!B) {
       continue;